@@ -1,53 +1,68 @@
 {
   "name": "integration-tests",
-  "author": "Zondax GmbH",
-  "license": "Apache-2.0",
   "version": "1.0.0",
   "description": "",
-  "types": "./dist/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Zondax/ledger-filecoin"
-  },
   "keywords": [
     "Zondax",
     "Ledger"
   ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Zondax/ledger-filecoin"
+  },
+  "license": "Apache-2.0",
+  "author": "Zondax AG",
+  "types": "./dist/index.d.ts",
   "scripts": {
-    "test": "ts-node tests/pullImageKillOld.ts && jest --maxConcurrency 2",
-    "try": "node try.mjs"
+    "clean": "ts-node tests/pullImageKillOld.ts",
+    "format": "FORCE_COLOR=1 prettier --write . && sort-package-json",
+    "format:check": "FORCE_COLOR=1 prettier --check .",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
+    "test": "ts-node tests/pullImageKillOld.ts && jest tests/raw.test.ts && jest --testPathIgnorePatterns 'tests/raw.test.ts' --maxConcurrency 2",
+    "try": "ts-node try.ts",
+    "upgrade": "bunx npm-check-updates -i"
+  },
+  "resolutions": {
+    "axios": "^1.10.0",
+    "glob": "^11.0.3"
   },
   "dependencies": {
-    "@ledgerhq/hw-transport-node-hid": "6.29.5",
-    "@zondax/ledger-filecoin": "^2.0.2",
-    "@zondax/zemu": "^0.53.0"
+    "@ledgerhq/hw-transport-node-hid": "6.29.8",
+    "@zondax/ledger-filecoin": "^3.0.0",
+    "@zondax/zemu": "^0.55.3",
+    "varint": "^6.0.0"
   },
   "devDependencies": {
-    "@ethereumjs/common": "^4.4.0",
-    "@ethereumjs/tx": "^5.4.0",
-    "@ledgerhq/hw-app-eth": "6.42.4",
+    "@ethereumjs/common": "^10.0.0",
+    "@ethereumjs/tx": "^10.0.0",
+    "@ledgerhq/hw-app-eth": "6.45.10",
     "@types/elliptic": "^6.4.18",
-    "@types/jest": "^29.5.14",
-    "@types/ledgerhq__hw-transport": "^4.21.8",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.0.14",
     "@types/secp256k1": "^4.0.6",
-    "@typescript-eslint/eslint-plugin": "^8.22.0",
-    "@typescript-eslint/parser": "^8.22.0",
+    "@typescript-eslint/eslint-plugin": "^8.37.0",
+    "@typescript-eslint/parser": "^8.37.0",
     "blakejs": "^1.2.1",
-    "bn.js": "^5.2.0",
+    "bn.js": "^5.2.2",
     "crypto-js": "4.2.0",
     "elliptic": "^6.6.1",
-    "eslint": "^9.19.0",
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-jest": "^28.11.0",
-    "eslint-plugin-prettier": "^5.2.3",
-    "jest": "29.7.0",
-    "jest-serial-runner": "^1.1.0",
-    "js-sha256": "0.11.0",
-    "prettier": "^3.4.2",
+    "eslint": "^9.31.0",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-jest": "^28.8.0",
+    "eslint-plugin-prettier": "^5.5.1",
+    "eslint-plugin-tsdoc": "^0.4.0",
+    "eslint-plugin-unused-imports": "^4.1.4",
+    "jest": "30.0.4",
+    "jest-runner": "^30.0.4",
+    "jest-serial-runner": "^1.2.2",
+    "js-sha256": "0.11.1",
+    "prettier": "^3.6.2",
     "secp256k1": "^5.0.1",
-    "ts-jest": "^29.2.5",
+    "sort-package-json": "^3.4.0",
+    "ts-jest": "^29.4.0",
     "ts-node": "^10.9.2",
-    "typescript": "^5.7.3"
+    "typescript": "^5.8.3"
   }
 }
