@@ -1,4 +1,4 @@
-// Modified by Zondax GmbH
+// Modified by Zondax AG
 //
 // Base32 implementation
 //
@@ -21,38 +21,42 @@
 
 #include <string.h>
 
-uint32_t base32_encode(const uint8_t *data,
-                       uint32_t length,
-                       char *result,
-                       uint32_t resultLen)
-{
-    if (data == NULL || result == NULL ||
-        length > (1 << 28) || length == 0 || resultLen == 0) {
+uint32_t base32_encode(const uint8_t *data, uint32_t length, char *result, uint32_t resultLen) {
+    static const char *alphabet = "abcdefghijklmnopqrstuvwxyz234567";
+
+    if (data == NULL || result == NULL || length == 0 || resultLen == 0 || length > (1u << 28)) {
         return 0;
     }
+
+    uint64_t buffer = 0;
+    uint32_t bitsLeft = 0;
     uint32_t count = 0;
-    uint32_t buffer = data[0];
-    uint32_t next = 1;
-    uint32_t bitsLeft = 8;
-    while (count < resultLen && (bitsLeft > 0 || next < length)) {
-        if (bitsLeft < 5) {
-            if (next < length) {
-                buffer <<= 8;
-                buffer |= data[next++] & 0xFF;
-                bitsLeft += 8;
-            } else {
-                uint32_t pad = 5u - bitsLeft;
-                buffer <<= pad;
-                bitsLeft += pad;
+
+    for (uint32_t i = 0; i < length; i++) {
+        buffer = (buffer << 8) | data[i];
+        bitsLeft += 8;
+
+        while (bitsLeft >= 5) {
+            if (count >= resultLen) {
+                return count;
             }
+            bitsLeft -= 5;
+            uint32_t index = (uint32_t)((buffer >> bitsLeft) & 0x1F);
+            result[count++] = alphabet[index];
         }
-        uint32_t index = 0x1Fu & (buffer >> (bitsLeft - 5u));
-        bitsLeft -= 5;
-        result[count++] = "abcdefghijklmnopqrstuvwxyz234567"[index];
+        // keep only the remaining bits in buffer to avoid overflow
+        buffer &= ((1ULL << bitsLeft) - 1);
+    }
+
+    // Handle the remaining bits (if any)
+    if (bitsLeft > 0 && count < resultLen) {
+        buffer <<= (5 - bitsLeft);
+        uint32_t index = (uint32_t)(buffer & 0x1F);
+        result[count++] = alphabet[index];
     }
 
     if (count < resultLen) {
-        result[count] = '\000';
+        result[count] = '\0';
     }
     return count;
 }
