@@ -1,26 +1,25 @@
 /*******************************************************************************
-*   (c) 2016 Ledger
-*   (c) 2018, 2019 Zondax GmbH
-*
-*  Licensed under the Apache License, Version 2.0 (the "License");
-*  you may not use this file except in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing, software
-*  distributed under the License is distributed on an "AS IS" BASIS,
-*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*  See the License for the specific language governing permissions and
-*  limitations under the License.
-********************************************************************************/
+ *   (c) 2016 Ledger
+ *   (c) 2018, 2019 Zondax AG
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ ********************************************************************************/
+#include <os_io_seproxyhal.h>
+
 #include "app_main.h"
 #include "view.h"
 
-#include <os_io_seproxyhal.h>
-
-__attribute__((section(".boot"))) int
-main(void) {
+__attribute__((section(".boot"))) int main(void) {
     // exit critical section
     __asm volatile("cpsie i");
 
@@ -28,14 +27,12 @@ main(void) {
     os_boot();
 
     BEGIN_TRY {
-        TRY{
+        TRY {
             app_init();
             app_main();
         }
-        CATCH_OTHER(e) {
-            UNUSED(e);
-        }
-        FINALLY{}
+        CATCH_OTHER(e) { UNUSED(e); }
+        FINALLY {}
     }
     END_TRY;
 }
