@@ -1,28 +1,30 @@
 /*******************************************************************************
-*  (c) 2019 Zondax GmbH
-*
-*  Licensed under the Apache License, Version 2.0 (the "License");
-*  you may not use this file except in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing, software
-*  distributed under the License is distributed on an "AS IS" BASIS,
-*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*  See the License for the specific language governing permissions and
-*  limitations under the License.
-********************************************************************************/
+ *  (c) 2019 Zondax AG
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ ********************************************************************************/
 #pragma once
 
-#include "coin.h"
-#include "zxerror.h"
 #include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
+
+#include "zxerror.h"
 
 // transaction initializer for the buffer and transaction type.
 void tx_context_fil();
-void tx_context_eth();
 void tx_context_raw_bytes();
+void tx_context_eth();
 
 // Signing is differently depending tx is rawBytes type
 bool tx_is_rawbytes();
@@ -59,9 +61,5 @@ const char *tx_parse(uint8_t *error_code);
 zxerr_t tx_getNumItems(uint8_t *num_items);
 
 /// Gets an specific item from the transaction (including paging)
-zxerr_t tx_getItem(int8_t displayIdx,
-                   char *outKey, uint16_t outKeyLen,
-                   char *outValue, uint16_t outValueLen,
+zxerr_t tx_getItem(int8_t displayIdx, char *outKey, uint16_t outKeyLen, char *outValue, uint16_t outValueLen,
                    uint8_t pageIdx, uint8_t *pageCount);
-
-zxerr_t tx_compute_eth_v(unsigned int info, uint8_t *v);
