@@ -1,193 +1,195 @@
 /*******************************************************************************
-*   (c) 2019 Zondax GmbH
-*
-*  Licensed under the Apache License, Version 2.0 (the "License");
-*  you may not use this file except in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing, software
-*  distributed under the License is distributed on an "AS IS" BASIS,
-*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*  See the License for the specific language governing permissions and
-*  limitations under the License.
-********************************************************************************/
+ *   (c) 2019 Zondax AG
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ ********************************************************************************/
 
-#include "gtest/gtest.h"
-#include <string>
 #include <cbor.h>
 #include <hexutils.h>
 #include <zxmacros.h>
 
+#include <string>
+
+#include "gtest/gtest.h"
+
 // Basic CBOR test cases generated with http://cbor.me/
 
 namespace {
-    TEST(CBORParserTest, MinimalListTest) {
-        // [1,2,3]
-        uint8_t inBuffer[100];
-        const char *tmp = "83010203";
-        auto inBufferLen = parseHexString(inBuffer, sizeof(inBuffer), tmp);
+TEST(CBORParserTest, MinimalListTest) {
+    // [1,2,3]
+    uint8_t inBuffer[100];
+    const char *tmp = "83010203";
+    auto inBufferLen = parseHexString(inBuffer, sizeof(inBuffer), tmp);
 
-        CborParser parser;
-        CborValue it;
-        CborError err;
+    CborParser parser;
+    CborValue it;
+    CborError err;
 
-        err = cbor_parser_init(inBuffer, inBufferLen, 0, &parser, &it);
-        EXPECT_EQ(err, CborNoError);
-        size_t arrLen;
-        err = cbor_value_get_array_length(&it, &arrLen);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(arrLen, 3);
+    err = cbor_parser_init(inBuffer, inBufferLen, 0, &parser, &it);
+    EXPECT_EQ(err, CborNoError);
+    size_t arrLen;
+    err = cbor_value_get_array_length(&it, &arrLen);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(arrLen, 3);
 
-        // Manually iterate
-        EXPECT_FALSE(cbor_value_at_end(&it));
+    // Manually iterate
+    EXPECT_FALSE(cbor_value_at_end(&it));
 
-        CborType type = cbor_value_get_type(&it);
-        EXPECT_EQ(type, CborArrayType);
-        EXPECT_TRUE(cbor_value_is_container(&it));
+    CborType type = cbor_value_get_type(&it);
+    EXPECT_EQ(type, CborArrayType);
+    EXPECT_TRUE(cbor_value_is_container(&it));
 
-        /// Enter container and iterate along items
-        CborValue contents;
-        err = cbor_value_enter_container(&it, &contents);
-        EXPECT_EQ(err, CborNoError);
+    /// Enter container and iterate along items
+    CborValue contents;
+    err = cbor_value_enter_container(&it, &contents);
+    EXPECT_EQ(err, CborNoError);
 
-        int64_t val;
-        // item = 1
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborIntegerType);
-        err = cbor_value_get_int64(&contents, &val);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(val, 1);
-        err = cbor_value_advance_fixed(&contents);
-        EXPECT_EQ(err, CborNoError);
+    int64_t val;
+    // item = 1
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborIntegerType);
+    err = cbor_value_get_int64(&contents, &val);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(val, 1);
+    err = cbor_value_advance_fixed(&contents);
+    EXPECT_EQ(err, CborNoError);
 
-        // item = 2
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborIntegerType);
-        err = cbor_value_get_int64(&contents, &val);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(val, 2);
-        err = cbor_value_advance_fixed(&contents);
-        EXPECT_EQ(err, CborNoError);
+    // item = 2
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborIntegerType);
+    err = cbor_value_get_int64(&contents, &val);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(val, 2);
+    err = cbor_value_advance_fixed(&contents);
+    EXPECT_EQ(err, CborNoError);
 
-        // item = 3
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborIntegerType);
-        err = cbor_value_get_int64(&contents, &val);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(val, 3);
-        err = cbor_value_advance_fixed(&contents);
-        EXPECT_EQ(err, CborNoError);
+    // item = 3
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborIntegerType);
+    err = cbor_value_get_int64(&contents, &val);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(val, 3);
+    err = cbor_value_advance_fixed(&contents);
+    EXPECT_EQ(err, CborNoError);
 
-        // Close container
-        err = cbor_value_leave_container(&it, &contents);
-        EXPECT_EQ(err, CborNoError);
-    }
-
-    TEST(CBORParserTest, MinimalDictTest) {
-        // { "x" : 1, "y" : 2, "z" : "test" }
-        uint8_t inBuffer[100];
-        const char *tmp = "A3617801617902617A6474657374";
-        auto inBufferLen = parseHexString(inBuffer, sizeof(inBuffer), tmp);
-
-        CborParser parser;
-        CborValue it;
-        CborError err = cbor_parser_init(inBuffer, inBufferLen, 0, &parser, &it);
-        EXPECT_EQ(err, CborNoError);
-
-        // Manually iterate
-        EXPECT_FALSE(cbor_value_at_end(&it));
-
-        CborType type = cbor_value_get_type(&it);
-        EXPECT_EQ(type, CborMapType);
-        EXPECT_TRUE(cbor_value_is_container(&it));
-        size_t mapLen;
-        err = cbor_value_get_map_length(&it, &mapLen);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(mapLen, 3);
-
-        /// Enter container and iterate along items
-        CborValue contents;
-        err = cbor_value_enter_container(&it, &contents);
-        EXPECT_EQ(err, CborNoError);
-
-        size_t key_len;
-        uint64_t val;
-        char buffer[100];
-        MEMZERO(buffer, 100);
-
-        // "x":1  key
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborTextStringType);
-        key_len = sizeof(buffer);
-        err = _cbor_value_copy_string(&contents, buffer, &key_len, nullptr);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(key_len, 1);
-        EXPECT_EQ(strlen((const char *) buffer), 1);
-        EXPECT_STREQ(buffer, "x");
-        err = cbor_value_advance(&contents);         // easier than advance_fixed, performance hit is small
-        EXPECT_EQ(err, CborNoError);
-
-        // "x":1  value
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborIntegerType);
-        err = cbor_value_get_uint64(&contents, &val);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(val, 1);
-        err = cbor_value_advance(&contents);         // easier than advance_fixed, performance hit is small
-        EXPECT_EQ(err, CborNoError);
-
-        // "y":2  key
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborTextStringType);
-        EXPECT_EQ(err, CborNoError);
-        key_len = sizeof(buffer);
-        err = _cbor_value_copy_string(&contents, buffer, &key_len, nullptr);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(key_len, 1);
-        EXPECT_EQ(strlen((const char *) buffer), 1);
-        EXPECT_STREQ(buffer, "y");
-        err = cbor_value_advance(&contents);         // easier than advance_fixed, performance hit is small
-        EXPECT_EQ(err, CborNoError);
-
-        // "y":2  value
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborIntegerType);
-        err = cbor_value_get_uint64(&contents, &val);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(val, 2);
-        err = cbor_value_advance(&contents);         // easier than advance_fixed, performance hit is small
-        EXPECT_EQ(err, CborNoError);
-
-        // "z":"test"  key
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborTextStringType);
-        EXPECT_EQ(err, CborNoError);
-        key_len = sizeof(buffer);
-        err = _cbor_value_copy_string(&contents, buffer, &key_len, nullptr);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(key_len, 1);
-        EXPECT_EQ(strlen((const char *) buffer), 1);
-        EXPECT_STREQ(buffer, "z");
-        err = cbor_value_advance(&contents);         // easier than advance_fixed, performance hit is small
-        EXPECT_EQ(err, CborNoError);
-
-        // "z":"test"  value
-        type = cbor_value_get_type(&contents);
-        EXPECT_EQ(type, CborTextStringType);
-        EXPECT_EQ(err, CborNoError);
-        key_len = sizeof(buffer);
-        err = _cbor_value_copy_string(&contents, buffer, &key_len, nullptr);
-        EXPECT_EQ(err, CborNoError);
-        EXPECT_EQ(key_len, 4);
-        EXPECT_EQ(strlen((const char *) buffer), 4);
-        EXPECT_STREQ(buffer, "test");
-        err = cbor_value_advance(&contents);         // easier than advance_fixed, performance hit is small
-        EXPECT_EQ(err, CborNoError);
-
-        // Close container
-        err = cbor_value_leave_container(&it, &contents);
-        EXPECT_EQ(err, CborNoError);
-    }
+    // Close container
+    err = cbor_value_leave_container(&it, &contents);
+    EXPECT_EQ(err, CborNoError);
 }
+
+TEST(CBORParserTest, MinimalDictTest) {
+    // { "x" : 1, "y" : 2, "z" : "test" }
+    uint8_t inBuffer[100];
+    const char *tmp = "A3617801617902617A6474657374";
+    auto inBufferLen = parseHexString(inBuffer, sizeof(inBuffer), tmp);
+
+    CborParser parser;
+    CborValue it;
+    CborError err = cbor_parser_init(inBuffer, inBufferLen, 0, &parser, &it);
+    EXPECT_EQ(err, CborNoError);
+
+    // Manually iterate
+    EXPECT_FALSE(cbor_value_at_end(&it));
+
+    CborType type = cbor_value_get_type(&it);
+    EXPECT_EQ(type, CborMapType);
+    EXPECT_TRUE(cbor_value_is_container(&it));
+    size_t mapLen;
+    err = cbor_value_get_map_length(&it, &mapLen);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(mapLen, 3);
+
+    /// Enter container and iterate along items
+    CborValue contents;
+    err = cbor_value_enter_container(&it, &contents);
+    EXPECT_EQ(err, CborNoError);
+
+    size_t key_len;
+    uint64_t val;
+    char buffer[100];
+    MEMZERO(buffer, 100);
+
+    // "x":1  key
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborTextStringType);
+    key_len = sizeof(buffer);
+    err = _cbor_value_copy_string(&contents, buffer, &key_len, nullptr);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(key_len, 1);
+    EXPECT_EQ(strlen((const char *)buffer), 1);
+    EXPECT_STREQ(buffer, "x");
+    err = cbor_value_advance(&contents);  // easier than advance_fixed, performance hit is small
+    EXPECT_EQ(err, CborNoError);
+
+    // "x":1  value
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborIntegerType);
+    err = cbor_value_get_uint64(&contents, &val);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(val, 1);
+    err = cbor_value_advance(&contents);  // easier than advance_fixed, performance hit is small
+    EXPECT_EQ(err, CborNoError);
+
+    // "y":2  key
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborTextStringType);
+    EXPECT_EQ(err, CborNoError);
+    key_len = sizeof(buffer);
+    err = _cbor_value_copy_string(&contents, buffer, &key_len, nullptr);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(key_len, 1);
+    EXPECT_EQ(strlen((const char *)buffer), 1);
+    EXPECT_STREQ(buffer, "y");
+    err = cbor_value_advance(&contents);  // easier than advance_fixed, performance hit is small
+    EXPECT_EQ(err, CborNoError);
+
+    // "y":2  value
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborIntegerType);
+    err = cbor_value_get_uint64(&contents, &val);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(val, 2);
+    err = cbor_value_advance(&contents);  // easier than advance_fixed, performance hit is small
+    EXPECT_EQ(err, CborNoError);
+
+    // "z":"test"  key
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborTextStringType);
+    EXPECT_EQ(err, CborNoError);
+    key_len = sizeof(buffer);
+    err = _cbor_value_copy_string(&contents, buffer, &key_len, nullptr);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(key_len, 1);
+    EXPECT_EQ(strlen((const char *)buffer), 1);
+    EXPECT_STREQ(buffer, "z");
+    err = cbor_value_advance(&contents);  // easier than advance_fixed, performance hit is small
+    EXPECT_EQ(err, CborNoError);
+
+    // "z":"test"  value
+    type = cbor_value_get_type(&contents);
+    EXPECT_EQ(type, CborTextStringType);
+    EXPECT_EQ(err, CborNoError);
+    key_len = sizeof(buffer);
+    err = _cbor_value_copy_string(&contents, buffer, &key_len, nullptr);
+    EXPECT_EQ(err, CborNoError);
+    EXPECT_EQ(key_len, 4);
+    EXPECT_EQ(strlen((const char *)buffer), 4);
+    EXPECT_STREQ(buffer, "test");
+    err = cbor_value_advance(&contents);  // easier than advance_fixed, performance hit is small
+    EXPECT_EQ(err, CborNoError);
+
+    // Close container
+    err = cbor_value_leave_container(&it, &contents);
+    EXPECT_EQ(err, CborNoError);
+}
+}  // namespace
