@@ -1,23 +1,23 @@
 /*******************************************************************************
-*   (c) 2019 Zondax GmbH
-*
-*  Licensed under the Apache License, Version 2.0 (the "License");
-*  you may not use this file except in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing, software
-*  distributed under the License is distributed on an "AS IS" BASIS,
-*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*  See the License for the specific language governing permissions and
-*  limitations under the License.
-********************************************************************************/
+ *   (c) 2019 Zondax AG
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ ********************************************************************************/
 #pragma once
 #include <fstream>
+#include <nlohmann/json.hpp>
 #include <vector>
-#include <json/json.h>
 
-std::vector<std::string> GenerateExpectedUIOutput(const Json::Value &json, bool expertMode);
-std::vector<std::string> EVMGenerateExpectedUIOutput(const Json::Value &json, bool expertMode);
-std::vector<std::string> InvokeContractGenerateExpectedUIOutput(const Json::Value &json, bool expertMode);
+std::vector<std::string> GenerateExpectedUIOutput(const nlohmann::json &json, bool expertMode);
+std::vector<std::string> EVMGenerateExpectedUIOutput(const nlohmann::json &json, bool expertMode);
+std::vector<std::string> InvokeContractGenerateExpectedUIOutput(const nlohmann::json &json, bool expertMode);
