@@ -1,10 +1,9 @@
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  transformIgnorePatterns: ['^.+\\.js$'],
-  globalSetup: "./jest/globalsetup.ts",
-  globalTeardown: "./jest/globalteardown.ts",
-  setupFilesAfterEnv: ["./jest/setup.ts"]
-}
-
-
+  preset: "ts-jest",
+  testEnvironment: "node",
+  transformIgnorePatterns: ["^.+\\.js$"],
+  reporters: ["default", ["summary", { summaryThreshold: 1 }]],
+  globalSetup: "./globalsetup.js",
+  globalTeardown: "./globalteardown.js",
+  testTimeout: 60000,
+};
