@@ -16,73 +16,73 @@
 
 import Zemu from "@zondax/zemu";
 // @ts-ignore
-import FilecoinApp from "@zondax/ledger-filecoin";
+import { FilecoinApp } from "@zondax/ledger-filecoin";
 import * as secp256k1 from "secp256k1";
-import { models, defaultOptions, PATH_TESTNET } from './common'
-import * as multisigData from "./multisig.json"
+import { models, defaultOptions, PATH_TESTNET } from "./common";
+import * as multisigData from "./multisig.json";
 import { getDigest } from "./utils";
 
 const TEST_DATA = [
   {
-    name: 'multisig_create',
+    name: "multisig_create",
     op: multisigData.create.cbor,
   },
   {
-    name: 'multisig_propose',
+    name: "multisig_propose",
     op: multisigData.propose.cbor,
   },
   {
-    name: 'multisig_approve',
+    name: "multisig_approve",
     op: multisigData.approve.cbor,
   },
   {
-    name: 'multisig_cancel',
+    name: "multisig_cancel",
     op: multisigData.cancel.cbor,
-  }
-]
+  },
+];
 
-jest.setTimeout(90000)
+jest.setTimeout(90000);
 
-describe.each(models)('Multisig', function (m) {
-  test.concurrent.each(TEST_DATA)('Multisig extended params', async function ({name, op}) {
-    const sim = new Zemu(m.path);
-    try {
-      await sim.start({...defaultOptions, model: m.name,});
-      const app = new FilecoinApp(sim.getTransport());
+describe.each(models)("Multisig", function (m) {
+  test.concurrent.each(TEST_DATA)(
+    "Multisig extended params",
+    async function ({ name, op }) {
+      const sim = new Zemu(m.path);
+      try {
+        await sim.start({ ...defaultOptions, model: m.name });
+        const app = new FilecoinApp(sim.getTransport());
 
-      const testcase = `${m.prefix.toLowerCase()}-sign-${name}`
+        const testcase = `${m.prefix.toLowerCase()}-sign-${name}`;
 
-      // Put the app in expert mode
-      await sim.toggleExpertMode();
+        // Put the app in expert mode
+        await sim.toggleExpertMode();
 
-      const txBlob = Buffer.from(op, 'hex');
+        const txBlob = Buffer.from(op, "hex");
 
-      const pkResponse = await app.getAddressAndPubKey(PATH_TESTNET);
-      console.log(pkResponse);
-      expect(pkResponse.return_code).toEqual(0x9000);
-      expect(pkResponse.error_message).toEqual("No errors");
+        const pkResponse = await app.getAddressAndPubKey(PATH_TESTNET);
+        console.log(pkResponse);
 
-      // do not wait here so we can get snapshots and interact with the app
-      const signatureRequest = app.sign(PATH_TESTNET, txBlob);
+        // do not wait here so we can get snapshots and interact with the app
+        const signatureRequest = app.sign(PATH_TESTNET, txBlob);
 
-      // Wait until we are not in the main menu
-      await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
-      await sim.compareSnapshotsAndApprove(".", testcase)
+        // Wait until we are not in the main menu
+        await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
+        await sim.compareSnapshotsAndApprove(".", testcase);
 
-      let resp = await signatureRequest;
-      console.log(resp, m.name, name)
+        let resp = await signatureRequest;
+        console.log(resp, m.name, name);
 
-      expect(resp.return_code).toEqual(0x9000);
-      expect(resp.error_message).toEqual("No errors");
-
-      // Verify signature
-      const pk = Uint8Array.from(pkResponse.compressed_pk)
-      const digest = getDigest(txBlob);
-      const signature = secp256k1.signatureImport(Uint8Array.from(resp.signature_der));
-      const signatureOk = secp256k1.ecdsaVerify(signature, digest, pk);
-      expect(signatureOk).toEqual(true);
-    } finally {
-      await sim.close();
-    }
-  });
-})
+        // Verify signature
+        const pk = Uint8Array.from(pkResponse.compressed_pk);
+        const digest = getDigest(txBlob);
+        const signature = secp256k1.signatureImport(
+          Uint8Array.from(resp.signature_der),
+        );
+        const signatureOk = secp256k1.ecdsaVerify(signature, digest, pk);
+        expect(signatureOk).toEqual(true);
+      } finally {
+        await sim.close();
+      }
+    },
+  );
+});
